requirements.txt
Django==1.8.3
django-ajaximage==0.2.8
django-annoying==0.9.0
django-apptemplates==0.0.1
django-braces==1.8.1
django-bulk-update==1.1.10
django-classy-tags==0.6.2
django-common-helpers==0.7.0
django-cors-headers==1.1.0
django-crispy-forms==1.4.0
django-cron==0.4.3
django-extensions==1.5.5
django-grappelli==2.7.1
django-grappelli-extensions==0.1.3
django-import-export==0.2.8
django-sslserver==0.15
djangorestframework==3.1.3
djangorestframework-camel-case==0.2.0
